third_party/WebKit/LayoutTests/fast/dom/custom/isolated-world.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<body>
<div id="container-a"><span is="x-a" id="x"></span></div>
<div id="container-b"><span is="x-a" id="y"></span></div>
<script>
description('Tests that accessing custom elements from an isolated world ' +
            'does not cause worlds to collide and destroy the galaxy.');

shouldBeNonNull('window.testRunner', 'this test requires testRunner');

var containerA = document.querySelector('#container-a');

// Access an upgrade candidate from an isolated world.
var world = 1;
testRunner.evaluateScriptInIsolatedWorld(
    world,
    'window.containerA = document.querySelector("#container-a"); ' +
    'window.a = containerA.firstElementChild; ' +
    'containerA.remove();');

// Registering an exception in the isolated world should raise an exception.
result = testRunner.evaluateScriptInIsolatedWorldAndReturnValue(
    world,
    '(function () { ' +
    '  var proto = Object.create(HTMLSpanElement.prototype); ' +
    '  proto.createdCallback = function () { ' +
    '    console.log(this.id + " entered (isolated)"); ' +
    '  }; ' +
    '  try { ' +
    '    document.registerElement("x-a", {extends: "span", prototype: proto}); ' +
    '    return "register succeeded"; ' +
    '  } catch (e) { ' +
    '    return e.code; ' +
    '  } ' +
    '})()');

shouldBe('result', 'DOMException.NOT_SUPPORTED_ERR',
         'calling register from an isolated world should throw an exception');

// Now access this element from the main world.
a = containerA.querySelector('#x');
shouldBe('a.__proto__', 'HTMLSpanElement.prototype',
         'the main world should see the usual main world prototype');

// Upgrade from the main world

var proto = Object.create(HTMLSpanElement.prototype);
invocations = [];
proto.attachedCallback = function () {
    invocations.push(this.id + ' entered (main)');
};
document.registerElement('x-a', {extends: 'span', prototype: proto});
shouldBe('invocations', '["y entered (main)"]',
         'only the element in the document should generate entered events');

// Insert from the isolated world

invocations = [];
testRunner.evaluateScriptInIsolatedWorld(world,
                                         'document.body.appendChild(a);');
shouldBe('invocations', '["x entered (main)"]',
         'modification in the isolated world should have caused callbacks ' +
         'in the main world');

// Examine prototypes from the isolated world

proto.p = 'p';
HTMLSpanElement.prototype.q = 'q';

result = testRunner.evaluateScriptInIsolatedWorldAndReturnValue(
    world,
    '(function () { ' +
    '  try { ' +
    '    var messages = []; ' +
    '    function log(message) { messages.push(message); } ' +
    '    function inspect(obj) { ' +
    '      log(obj.id + " prototype is HTMLSpanElement.prototype? " + ' +
    '          (obj.__proto__ === HTMLSpanElement.prototype)); ' +
    '      log(obj.id + ".p ~> " + obj.p); ' +
    '      log(obj.id + ".q ~> " + obj.q); ' +
    '    } ' +
    '    inspect(window.a); ' +
    '    inspect(document.querySelector("#y")); ' +
    '    return messages.join(", "); ' +
    '  } catch (e) { ' +
    '    return e.toString(); ' +
    '  } ' +
    '})()');
shouldBe(
    'result',
    '"x prototype is HTMLSpanElement.prototype? true,' +
    ' x.p ~> undefined,' +
    ' x.q ~> undefined,' +
    ' y prototype is HTMLSpanElement.prototype? true,' +
    ' y.p ~> undefined,' +
    ' y.q ~> undefined"',
    'the isolated world should not see main world prototypes');

// Lastly, collect some wrappers (e.g #b in isolated world)
document.querySelector('#container-b').innerHTML = '';
gc();
testPassed('did not crash');
</script>
